Make it easier to discover the (excellent) developer docs
As mentioned in #253, it's a real shame to have such great developer
docs which are "hidden" by not being conspicuously linked from
anywhere.

So link to them from the top-level README.md, and also add a
CONTRIBUTING.md which is a conventional file location understood by
sites such as GitHub:

  https://github.com/open-source-labs/reactime/community
  https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors#adding-a-contributing-file

Partially addresses #253.